Fix adopted records being wrong colour
When an individual is adopted, they are not a blood relative (unless also related in a way other than by adoption). This change fixes the colouring of adopted records by updating existing non-functioning code to work with recent versions of webtrees, and then turning this into a function to be called from various places affected.
Resolves #36