Fix related colours when starting from adopted individual
If adopted person was chosen as starting individual, siblings were assumed as related and this spread through the tree. Fixed by adding check for if the individual being added is not related to their own family, if they aren't related then set $rel to false for this round of recursion.