Fix combined mode with stop indi
In combined mode, if a stop individual is selected it is possible a record can be left incomplete. This commit adds a scan to remove family records where there is no spouse or children in combined mode - all records should have a spouse in combined mode. Dummy families (where no spouse exists) are ignored.