Add error when canvas too big
If exporting file uses canvas (e.g. JPEG, PNG) then the browser has a max canvas size that varies by browser and in most cases will result in a blank canvas and an invalid zero kilobyte file downloaded. If the output is very short we have hit this limit, so we show an error message.