Set GV_FILE_PATH
In hope of fixing issue, have set GV_FILE_PATH to /

Note recent versions of Graphviz don't use this, so setting it to / should be equivalent to not having it.

Can't be more specific as needs to access both Webtrees media storage and temporary directory, and I can't find any documentation nor can I reproduce the issue.